+++ RNMultiPicker.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:VDRPickerIOS.xcodeproj">
+      location = "self:RNMultiPicker.xcodeproj">
    </FileRef>
 </Workspace>
