+++ RNMultiPicker/RNMultiPicker.h
@@ -1,7 +1,4 @@
 //
-//  VDRPickerView.h
-//  VedderPicker
-//
 //  Created by David Vedder on 4/15/15.
 //  Copyright (c) 2015 Facebook. All rights reserved.
 //
@@ -9,7 +6,7 @@
 #import <UIKit/UIKit.h>
 @class RCTEventDispatcher;
 
-@interface VDRPicker : UIPickerView
+@interface RNMultiPicker : UIPickerView
 
 - (instancetype)initWithEventDispatcher:(RCTEventDispatcher *)eventDispatcher NS_DESIGNATED_INITIALIZER;
