+++ RNMultiPicker/RNMultiPickerManager.h
@@ -1,13 +1,10 @@
 //
-//  VDRPickerManager.h
-//  VedderPicker
-//
 //  Created by David Vedder on 4/16/15.
 //  Copyright (c) 2015 Facebook. All rights reserved.
 //
 
 #import "RCTViewManager.h"
 
-@interface VDRPickerManager : RCTViewManager
+@interface RNMultiPickerManager : RCTViewManager
 
 @end
